Fix in 'johnny/tests/base.py#supports_transactions' function.
Function was broken since con.features.__dict__.get() isn't guaranteed that it will get real feature value. For example, 'django.db.backends.BaseDatabaseFeatures#supports_transactions' is implemented as a 'cached_property'. The 'cached_property' decorator is implemented as a descriptor that replaces itself on the instance with computed value after first call - it means that prior to first call instance it is attached to has no value for target field.

TL;DR: Added support for feature fields based on descriptors.